refactor first done
