+++ server/meep-draft/defaultSettings/default-controls.js
@@ -1,0 +1,30 @@
+const defaultControls = {
+  fontFamily: true,
+  fontSize: true,
+  text: {
+    BOLD: true,
+    ITALIC: true,
+    UNDERLINE: true,
+    STRIKETHROUGH: true
+  },
+  link: {
+    set: true,
+    unset: true
+  },
+  block: {
+    headerTwo: true,
+    unorderedListItem: true,
+    orderedListItem: true,
+    alignLeft: true,
+    alignCenter: true,
+    alignRight: true
+  },
+  color: true,
+  background: true,
+  content: {
+    undo: true,
+    redo: true
+  }
+}
+
+export default defaultControls;
